requirements.txt
# Python >=3.11,<3.13
Pillow>=10.0.0,<13.0.0
pillow-heif>=1.0.0
